refactor: Move user management to separate page
